fix: move ProviderType to shared types and update usages (#7710)
(cherry picked from commit 60e0040577a452224dca5c18e4cd33cccc46ddd9)

# Conflicts:
#	ui/CHANGELOG.md
