+++ ui/components/providers/workflow/provider-title-docs.tsx
@@ -1,8 +1,8 @@
 import Link from "next/link";
 
 import { getProviderName } from "@/components/ui/entities/get-provider-logo";
-import { ProviderType } from "@/components/ui/entities/get-provider-logo";
 import { getProviderLogo } from "@/components/ui/entities/get-provider-logo";
+import { ProviderType } from "@/types";
 
 export const ProviderTitleDocs = ({
   providerType,
