+++ ui/types/components.ts
@@ -1,6 +1,8 @@
 import { LucideIcon } from "lucide-react";
 import { SVGProps } from "react";
 
+import { ProviderType } from "./providers";
+
 export type IconSvgProps = SVGProps<SVGSVGElement> & {
   size?: number;
 };
@@ -45,7 +47,7 @@ export interface CollapseMenuButtonProps {
 export interface SelectScanComplianceDataProps {
   scans: (ScanProps & {
     providerInfo: {
-      provider: "aws" | "azure" | "gcp" | "kubernetes";
+      provider: ProviderType;
       uid: string;
       alias: string;
     };
@@ -169,27 +171,6 @@ export interface FindingsSeverityOverview {
       informational: number;
     };
   };
-  meta: {
-    version: string;
-  };
-}
-
-export interface ProviderOverviewProps {
-  data: {
-    type: "provider-overviews";
-    id: "aws" | "gcp" | "azure" | "kubernetes";
-    attributes: {
-      findings: {
-        pass: number;
-        fail: number;
-        manual: number;
-        total: number;
-      };
-      resources: {
-        total: number;
-      };
-    };
-  }[];
   meta: {
     version: string;
   };
@@ -509,51 +490,6 @@ export interface UserProps {
   }[];
 }
 
-export interface ProviderProps {
-  id: string;
-  type: "providers";
-  attributes: {
-    provider: "aws" | "azure" | "m365" | "gcp" | "kubernetes";
-    uid: string;
-    alias: string;
-    status: "completed" | "pending" | "cancelled";
-    resources: number;
-    connection: {
-      connected: boolean;
-      last_checked_at: string;
-    };
-    scanner_args: {
-      only_logs: boolean;
-      excluded_checks: string[];
-      aws_retries_max_attempts: number;
-    };
-    inserted_at: string;
-    updated_at: string;
-    created_by: {
-      object: string;
-      id: string;
-    };
-  };
-  relationships: {
-    secret: {
-      data: {
-        type: string;
-        id: string;
-      } | null;
-    };
-    provider_groups: {
-      meta: {
-        count: number;
-      };
-      data: Array<{
-        type: string;
-        id: string;
-      }>;
-    };
-  };
-  groupNames?: string[];
-}
-
 export interface ScanProps {
   type: "scans";
   id: string;
@@ -596,7 +532,7 @@ export interface ScanProps {
     };
   };
   providerInfo?: {
-    provider: "aws" | "azure" | "gcp" | "kubernetes";
+    provider: ProviderType;
     uid: string;
     alias: string;
   };
