+++ ui/types/index.ts
@@ -2,3 +2,4 @@ export * from "./authFormSchema";
 export * from "./components";
 export * from "./filters";
 export * from "./formSchemas";
+export * from "./providers";
